+++ hrk_nasa_apis/lib/src/nasa.dart
@@ -1,0 +1,3 @@
+class Nasa {
+  static final Uri url = Uri.https('nasa.gov');
+}