+++ hrk_nasa_apis/lib/src/nasa_apis.dart
@@ -1,7 +1,11 @@
+import 'nasa.dart';
+
 export 'globals.dart';
 export 'jpl_nasa.dart';
+export 'nasa.dart';
 export 'services/services.dart';
 
 class NasaApis {
-  static const String domain = 'nasa.gov';
+  static final Uri url = Uri.https('api.${Nasa.url.authority}');
+  static final Uri docUrl = url;
 }