+++ hrk_nasa_apis/lib/src/services/ssd_cneos/sbdb_cad_api/sbdb_cad_api.dart
@@ -19,11 +19,8 @@ class SbdbCadApi {
       ..transformer = SbdbCadTransformer();
   }
 
-  static final Uri url = Uri(
-    scheme: SsdCneos.baseUrl.scheme,
-    host: SsdCneos.baseUrl.host,
-    path: 'cad.api',
-  );
+  static final Uri url = Uri.https(SsdCneos.url.authority, 'cad.api');
+  static final Uri docUrl = Uri.https(SsdCneos.url.authority, 'doc/cad.html');
   static const String displayName = 'SBDB Close-Approach Data API';
   static const String version = '1.5';
   late final Dio _dio;