+++ hrk_nasa_apis/lib/src/services/ssd_cneos/ssd_cneos.dart
@@ -5,11 +5,8 @@ export 'data/data.dart';
 export 'sbdb_cad_api/sbdb_cad_api.dart';
 
 class SsdCneos {
-  static final Uri baseUrl = Uri(
-    scheme: 'https',
-    host: 'ssd-api.${JplNasa.domain}',
-  );
+  static final Uri url = Uri.https('ssd-api.${JplNasa.url.authority}');
   static final BaseOptions dioBaseOptions = BaseOptions(
-    baseUrl: baseUrl.toString(),
+    baseUrl: url.toString(),
   );
 }